requirements.txt
pytorch-lightning
pytest
pytest-cov
torchvision
sphinx
myst-parser
furo
pycryptodome